Cmake 2.6.0 is now required. Moved Macro to separate file to be able to use it in more project than those that depend on H3DUtil. Added macro to check for vcexpress version used FindDirectShow.cmake. Made some cache variables into advanced variables in modules. Fixed FindHAPI.cmake so that it behaves like FindH3DUtil and FindH3DAPI. Added FindGLUTWin used by h3dload and examples c++ demos.
Fixed some compile warnings due to conversion in some source files.
Fixed bugs in SpaceballSensor.cpp when it resizes mfbool Field.
Fixed "bug" in SplineScalarInterpolator.cpp where "true" was used instead of true.

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@991 610dfd11-2700-0410-9b68-9ae28eac5983

+++ include/H3D/OggFileReader.h
@@ -80,7 +80,7 @@ namespace H3D {
     
     /// Returns the total size of the PCM data of the current stream.
     virtual unsigned int totalDataSize() {
-      return ov_raw_total( &ogg_file, -1 );
+      return (unsigned int)( ov_raw_total( &ogg_file, -1 ) );
     }
       
     /// Returns the number of channels per second for the current PCM
