+++ src/H3DImageObject.cpp
@@ -40,11 +40,11 @@ void H3DImageObject::SFImage::setPixel( unsigned int x, unsigned int y,
       y_min = y_max = y;
       z_min = z_max = 0;
     } else {
-      if( x < x_min ) x_min = x;
-      else if( x > x_max ) x_max = x;
+      if( x < (unsigned int)x_min ) x_min = x;
+      else if( x > (unsigned int)x_max ) x_max = x;
     
-      if( y < y_min ) y_min = y;
-      else if( y > y_max ) y_max = y;
+      if( y < (unsigned int)y_min ) y_min = y;
+      else if( y > (unsigned int)y_max ) y_max = y;
     }
     
     value->setPixel( color, x, y );
@@ -54,8 +54,8 @@ void H3DImageObject::SFImage::setPixel( unsigned int x, unsigned int y,
 void H3DImageObject::SFImage::setPixel( const Vec2f &tc,
                                           const RGBA &color ) {
   if( value.get() ) {
-    setPixel( (value->width() - 1) * tc.x, 
-              (value->height() - 1) * tc.y,
+    setPixel( (unsigned int)( (value->width() - 1) * tc.x ),
+              (unsigned int)( (value->height() - 1) * tc.y ),
                color );
   }
 }
@@ -70,14 +70,14 @@ void H3DImageObject::SFImage::setPixel( unsigned int x,
       y_min = y_max = y;
       z_min = z_max = z;
     } else {
-      if( x < x_min ) x_min = x;
-      else if( x > x_max ) x_max = x;
+      if( x < (unsigned int)x_min ) x_min = x;
+      else if( x > (unsigned int)x_max ) x_max = x;
     
-      if( y < y_min ) y_min = y;
-      else if( y > y_max ) y_max = y;
+      if( y < (unsigned int)y_min ) y_min = y;
+      else if( y > (unsigned int)y_max ) y_max = y;
 
-      if( z < z_min ) z_min = z;
-      else if( z > z_max ) z_max = z;
+      if( z < (unsigned int)z_min ) z_min = z;
+      else if( z > (unsigned int)z_max ) z_max = z;
     }
     
     value->setPixel( color, x, y, z );
@@ -87,9 +87,9 @@ void H3DImageObject::SFImage::setPixel( unsigned int x,
 void H3DImageObject::SFImage::setPixel( const Vec3f &tc,
                                         const RGBA &color ) {
   if( value.get() ) {
-    setPixel( (value->width() - 1) * tc.x, 
-              (value->height() - 1) * tc.y,
-              (value->depth() - 1) * tc.z,
+    setPixel( (unsigned int)( (value->width() - 1) * tc.x ),
+              (unsigned int)( (value->height() - 1) * tc.y ),
+              (unsigned int)( (value->depth() - 1) * tc.z ),
                color );
   }
 }
